+++ Specs/DateTime/English/DateTimeModel.json
@@ -24275,7 +24275,6 @@
   },
   {
     "Input": "I'll be back 7:56:30 am",
-    "Comment": "Taskmode changes resolution of past time resolution to datetime.",
     "Context": {
       "ReferenceDateTime": "2016-11-07T08:00:00"
     },
